Make tile URL path configurable
Until now the tile URL path was hardcoded. But because we want to allow
arbitrary map style parameters (initially to support a language setting, but
other parameters should be possible), we need to make this configurable. To
achieve this the hard coded path parsing using boost::spirit is replaced
by a regular expression parsing using boost::xpressive that can be configured
at runtime. A new configuration parameter tile_path_template in the [mongrel2]
section of the tile_handler.conf can be used to configure the path.