Allowing to modify queuing priority based on style
To do this a new attribute "priority" was added to the tile_protocol class and
the protocol buffers tile. Instead of calculating the priority from the command
the broker now calls get_priority() on the tile. If no priority was set, this
will fall back to the old behaviour. But tile_handler normally sets the priority
now based on the command (as before) plus a value from the config file that
depends on the style. This way we can prioritize rendering certain styles.