Even hackier code to do multilingual labels.
Invented a new parameter for mapnik datasources called "labelhint" that is used
to decide whether and how labels should be rendered. This commit contains a
little bit of hackery in the mapnik worker to do all of this. This is very
specific to the multilingual map demo and needs some refactoring to separate
the general stuff from the specific stuff.