Move Mapnik singleton class in here.
We were using a singleton template class from Mapnik in a few places.
Unfortunately the interface changed in Mapnik at some point. To make our code
here work with both old and new Mapnik versions it seemed easiest just to pull
in the class as used in Mapnik into our code here. This is not the cleanest way
of doing it, especially because there is still a dependency on the
MAPNIK_THREADSAFE macro. So some further cleanup is necessary at some point.